backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1
apscheduler==3.10.4
google-generativeai==0.8.3
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
aiosqlite==0.20.0
bcrypt==4.2.1
